refactor health service package
