+++ services/src/main/proto/health.proto
@@ -31,7 +31,10 @@ syntax = "proto3";
 
 package grpc.health.v1;
 option csharp_namespace = "Grpc.Health.V1";
-option java_package = "io.grpc.services.health";
+
+option java_multiple_files = true;
+option java_package = "io.grpc.health.v1";
+option java_outer_classname = "HealthProto";
 
 message HealthCheckRequest {
   string service = 1;
